move safe_extract to h5io
